refactor : 네이밍 및 코드 간소화
코드 리뷰 피드백 바탕으로 파일 네이밍 변경 및 불필요한 코드 간소화
