+++ src/functions/reduceDataFunc.ts
@@ -1,4 +1,4 @@
-import { dataArray, dataObject, data } from "./funcData";
+import { dataArray, dataObject, data, dataColor } from "./funcData";
 
 // 데이터 내 지역 이름을 추출하여 중복없는 지역 이름 배열을 return하는 함수
 export const reduceById = () => {
@@ -38,52 +38,28 @@ export const reduceAreaData = () => {
 
 // 데이터 ID별 BackGroundColor 배열 return하는 함수
 export const reduceDataColor = () => {
-  const dataColor = Object.values(dataObject).reduce((acc: string[], cur) => {
-    switch (cur.id) {
-      case "성북구":
-        acc.push("#204051");
-        break;
-      case "강남구":
-        acc.push("#3B6978");
-        break;
-      case "노원구":
-        acc.push("#84A9AC");
-        break;
-      case "중랑구":
-        acc.push("#CAE8D5");
-        break;
-      default:
-        break;
-    }
-    return acc;
-  }, []);
-  return dataColor;
+  const dataColorArray = Object.values(dataObject).reduce(
+    (acc: string[], cur) => {
+      acc.push(dataColor[cur.id]);
+      return acc;
+    },
+    []
+  );
+  return dataColorArray;
 };
 
 // params.id와 data ID의 일치 여부에 따라 BackGroundColor 배열을 return 하는 함수
 export const reduceSelectedDataColor = (id: string) => {
-  const dataColor = Object.values(dataObject).reduce((acc: string[], cur) => {
-    if (cur.id !== id) {
-      acc.push("lightgray");
-    } else {
-      switch (cur.id) {
-        case "성북구":
-          acc.push("#204051");
-          break;
-        case "강남구":
-          acc.push("#3B6978");
-          break;
-        case "노원구":
-          acc.push("#84A9AC");
-          break;
-        case "중랑구":
-          acc.push("#CAE8D5");
-          break;
-        default:
-          break;
+  const dataColorArray = Object.values(dataObject).reduce(
+    (acc: string[], cur) => {
+      if (cur.id !== id) {
+        acc.push("lightgray");
+      } else {
+        acc.push(dataColor[cur.id]);
       }
-    }
-    return acc;
-  }, []);
-  return dataColor;
+      return acc;
+    },
+    []
+  );
+  return dataColorArray;
 };
